requirements.txt
absl-py==0.13.0
actionlib==1.14.0
angles==1.9.13
apturl==0.5.2
astunparse==1.6.3
atomicwrites==1.1.5
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==4.2.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.6
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.21
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.3
empy==3.3.2
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
fastimport==0.9.8
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gdown==3.13.1
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg==1.13.1
gps==3.20
grpcio==1.32.0
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib-metadata==1.5.0
incremental==16.10.1
interactive-markers==1.12.0
jdcal==1.0
jetson-stats==4.2.1
Jetson.GPIO==2.0.17
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Markdown==3.3.4
matplotlib==3.1.2
message-filters==1.16.0
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
netifaces==0.10.4
nose==1.3.7
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
openpyxl==3.0.3
opt-einsum==3.3.0
packaging==20.3
pandas==0.25.3
paramiko==2.6.0
pbr==5.6.0
pexpect==4.6.0
Pillow==9.2.0
pluggy==0.13.0
protobuf==3.18.0
psutil==5.5.1
py==1.8.1
py-ubjson==0.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylibi2c==0.2.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.4
PySocks==1.7.1
pytest==4.6.9
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.22.2
rosdep-modules==0.22.2
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosgraph==1.16.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospkg-modules==1.5.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.7.2
rviz==1.14.20
scipy==1.3.3
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.15.0
smach==2.5.2
smach-ros==2.5.2
smbus2==0.4.2
smclib==1.8.6
soupsieve==1.9.5
ssh-import-id==5.10
system-service==0.3
systemd-python==234
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorrt==8.0.1.6
termcolor==1.1.0
testresources==2.0.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
topic-tools==1.16.0
torch==1.13.0a0+git7c98e70
torchvision==0.14.0a0+5ce4506
tqdm==4.62.3
Twisted==18.9.0
txaio==2.10.0
typing-extensions==3.7.4.3
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.8
urwid==2.0.1
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.16
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
youtube-dl==2020.3.24
zipp==1.0.0
zope.interface==4.7.1
